python/requirements.txt
pyaudio
webrtcvad
openai-whisper
numpy
noisereduce